Added EJS support, developed file structure more
