+++ views/missions/show.ejs
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>Show Page</title>
+  </head>
+  <body>
+    <h1>Show Page</h1>
+  </body>
+</html>
